requirements.txt
boto3
IPy
dnspython
paramiko
progressbar
progressbar2